cleanup of schema explorer module
