+++ data/validation_schemas/class_json_schema.json
