+++ data/validation_schemas/property_json_schema.json
