+++ ingresspipe/schemas/__init__.py
@@ -1,2 +1,3 @@
-from .explorer import SchemaValidator, SchemaExplorer
+from .explorer import SchemaExplorer
 from .generator import SchemaGenerator
+from .validator import SchemaValidator
