+++ ingresspipe/schemas/explorer.py
@@ -8,156 +8,10 @@ import networkx as nx
 from networkx.algorithms.cycles import find_cycle
 
 from .base import *
-from .utils import *
+from ingresspipe.utils.general import expand_curies_in_schema, uri2label, find_duplicates
 from .curie import uri2curie, curie2uri
 
-_ROOT = os.path.abspath(os.path.dirname(__file__))
 namespaces = dict(rdf=Namespace("http://www.w3.org/1999/02/22-rdf-syntax-ns#"))
-
-class SchemaValidator():
-    """Validate Schema against SchemaOrg standard
-
-    Validation Criterias:
-    1. Data Structure wise:
-      > "@id", "@context", "@graph"
-      > Each element in "@graph" should contain "@id", "@type", "rdfs:comment",
-      "rdfs:label", "sms:displayName"
-      > validate against JSON Schema
-      > Should validate the whole structure, and also validate property and 
-      value separately
-    2. Data Content wise:
-      > "@id" field should match with "rdfs:label" field
-      > all prefixes used in the file should be defined in "@context"
-      > There should be no duplicate "@id"
-      > Class specific
-        > rdfs:label field should be capitalize the first character of each 
-          word for a class; 
-        > the value of "rdfs:subClassOf" should be present in the schema or in 
-          the core vocabulary
-        > sms:displayName ideally should contain capitalized words separated by space, but that's not enforced by validation
-      > Property specific
-        > rdfs:label field should be cammelCase
-        > the value of "schema:domainIncludes" should be present in the schema 
-          or in the core vocabulary
-        > the value of "schema:rangeIncludes" should be present in the schema 
-          or in the core vocabulary
-        > sms:displayName ideally should contain capitalized words separated by space, but that's not enforced by validation
-        TODO: add dependencies and component dependencies to class structure documentation; as well as value range and required property
-      
-    """
-    def __init__(self, schema):
-        self.schemaorg = {'schema': load_schemaorg(),
-                          'classes': [],
-                          'properties': []}
-        for _schema in self.schemaorg['schema']['@graph']:
-            for _record in _schema["@graph"]:
-                if "@type" in _record:
-                    _type = str2list(_record["@type"])
-                    if "rdfs:Property" in _type:
-                        self.schemaorg['properties'].append(_record["@id"])
-                    elif "rdfs:Class" in _type:
-                        self.schemaorg['classes'].append(_record["@id"])
-        self.extension_schema = {'schema': expand_curies_in_schema(schema),
-                                 'classes': [],
-                                 'properties': []}
-        for _record in self.extension_schema['schema']["@graph"]:
-            _type = str2list(_record["@type"])
-            if "rdfs:Property" in _type:
-                self.extension_schema['properties'].append(_record["@id"])
-            elif "rdfs:Class" in _type:
-                self.extension_schema['classes'].append(_record["@id"])
-        self.all_classes = self.schemaorg['classes'] + self.extension_schema['classes']
-
-    def validate_class_label(self, label_uri):
-        """ Check if the first character of class label is capitalized
-        """
-        label = extract_name_from_uri_or_curie(label_uri)
-        assert label[0].isupper()
-
-    def validate_property_label(self, label_uri):
-        """ Check if the first character of property label is lower case
-        """
-        label = extract_name_from_uri_or_curie(label_uri)
-        assert label[0].islower()
-
-    def validate_subclassof_field(self, subclassof_value):
-        """ Check if the value of "subclassof" is included in the schema file
-        """
-        subclassof_value = dict2list(subclassof_value)
-        for record in subclassof_value:
-            assert record["@id"] in self.all_classes
-
-    def validate_domainIncludes_field(self, domainincludes_value):
-        """ Check if the value of "domainincludes" is included in the schema
-        file
-        """
-        domainincludes_value = dict2list(domainincludes_value)
-        for record in domainincludes_value:
-            assert record["@id"] in self.all_classes, "value of domainincludes not recorded in schema: %r" % domainincludes_value
-
-    def validate_rangeIncludes_field(self, rangeincludes_value):
-        """ Check if the value of "rangeincludes" is included in the schema
-        file
-        """
-        rangeincludes_value = dict2list(rangeincludes_value)
-        for record in rangeincludes_value:
-            assert record["@id"] in self.all_classes
-
-    def check_whether_atid_and_label_match(self, record):
-        """ Check if @id field matches with the "rdfs:label" field
-        """
-        _id = extract_name_from_uri_or_curie(record["@id"])
-        assert _id == record["rdfs:label"], "id and label not match: %r" % record
-
-    def check_duplicate_labels(self):
-        """ Check for duplication in the schema
-        """
-        labels = [_record['rdfs:label'] for _record in self.extension_schema["schema"]["@graph"]]
-        duplicates = find_duplicates(labels)
-        try:
-            assert len(duplicates) == 0
-        except:
-            raise Exception('Duplicates detected in graph: ', duplicates)
-
-    def validate_schema(self, schema):
-        """Validate schema against SchemaORG standard
-        """
-        json_schema_path = os.path.join(_ROOT, 'data', 'schema.json')
-        json_schema = load_json(json_schema_path)
-        return validate(schema, json_schema)
-
-    def validate_property_schema(self, schema):
-        """Validate schema against SchemaORG property definition standard
-        """
-        json_schema_path = os.path.join(_ROOT,
-                                        'data',
-                                        'property_json_schema.json')
-        json_schema = load_json(json_schema_path)
-        return validate(schema, json_schema)
-
-    def validate_class_schema(self, schema):
-        """Validate schema against SchemaORG class definition standard
-        """
-        json_schema_path = os.path.join(_ROOT,
-                                        'data',
-                                        'class_json_schema.json')
-        json_schema = load_json(json_schema_path)
-        return validate(schema, json_schema)
-
-    def validate_full_schema(self):
-        self.check_duplicate_labels()
-        for record in self.extension_schema['schema']['@graph']:
-            self.check_whether_atid_and_label_match(record)
-            if record['@type'] == "rdf:Class":
-                self.validate_class_schema(record)
-                self.validate_class_label(record["@id"])
-            elif record['@type'] == "rdf:Property":
-                self.validate_property_schema(record)
-                self.validate_property_label(record["@id"])
-                self.validate_domainIncludes_field(record["http://schema.org/domainIncludes"])
-                if "http://schema.org/rangeIncludes" in record:
-                    self.validate_rangeIncludes_field(record["http://schema.org/rangeIncludes"])
-
 
 class SchemaExplorer():
     """Class for exploring schema
@@ -169,7 +23,6 @@ class SchemaExplorer():
         """Load schema and convert it to networkx graph
         """
         self.schema = load_json(schema)
-        validate_schema(self.schema)
         self.schema_nx = load_schema_into_networkx(self.schema)
 
     def load_default_schema(self):
@@ -507,9 +360,6 @@ class SchemaExplorer():
         """Edit an existing class into schema
         """ 
         for i, schema_class in enumerate(self.schema["@graph"]):
-            print(schema_class)
-            print(schema_class["rdfs:label"])
-            print(class_info["rdfs:label"])
             if schema_class["rdfs:label"] == class_info["rdfs:label"]:
                 validate_class_schema(class_info)
