--- ingresspipe/schemas/tests/data/property_schema_missing_domain.json
@@ -1,12 +1,0 @@
-{
-    "@id": "bts:precedes",
-    "@type": "rdf:Property",
-    "rdfs:comment": "holds between two processes, where one completes before the other begins",
-    "rdfs:label": "precedes",
-    "schema:isPartOf": {
-        "@id": "http://schema.biothings.io"
-    },
-    "schema:rangeIncludes": {
-        "@id": "bts:Occurrent"
-    }
-}
