+++ ingresspipe/utils/general.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Optional, Text
 import pandas as pd
 
 def update_df(existing_df:pd.DataFrame, new_df:pd.DataFrame, idx_key:str) -> pd.DataFrame:
-    """ Updates an existing data frame with the entries of another dataframe on a given primary index
+    """ Updates an existing data frame with the entries of another dataframe on a given primary index.
 
     Args: 
         existing_df : data frame to be updated
@@ -36,7 +36,8 @@ def update_df(existing_df:pd.DataFrame, new_df:pd.DataFrame, idx_key:str) -> pd.
 
 def execute_google_api_requests(service, requests_body, **kwargs):
     """
-    Execute google API requests batch; attempt to execute in parallel
+    Execute google API requests batch; attempt to execute in parallel.
+
     Args:
         service: google api service; for now assume google sheets service that is instantiated and authorized
         service_type: default batchUpdate; TODO: add logic for values update
@@ -45,9 +46,72 @@ def execute_google_api_requests(service, requests_body, **kwargs):
     """
 
     if "spreadsheet_id" in kwargs and "service_type" in kwargs and kwargs["service_type"] == "batch_update":
-
         # execute all requests
-
         response = service.spreadsheets().batchUpdate(spreadsheetId=kwargs["spreadsheet_id"], body = requests_body).execute()
         
         return response
+
+
+# utilities for schema explorer methods
+
+def expand_curie_to_uri(curie, context_info):
+    """Expand curie to uri based on the context given
+
+    parmas
+    ======
+    curie: curie to be expanded (e.g. bts:BiologicalEntity)
+    context_info: jsonld context specifying prefix-uri relation (e.g. {"bts":
+    "http://schema.biothings.io/"})
+    """
+    # as suggested in SchemaOrg standard file, these prefixes don't expand
+    PREFIXES_NOT_EXPAND = ["rdf", "rdfs", "xsd"]
+    # determine if a value is curie
+    if len(curie.split(':')) == 2:
+        prefix, value = curie.split(":")
+        if prefix in context_info and prefix not in PREFIXES_NOT_EXPAND:
+            return context_info[prefix] + value
+    # if the input is not curie, return the input unmodified
+        else:
+            return curie
+    else:
+        return curie
+
+
+def expand_curies_in_schema(schema):
+    """Expand all curies in a SchemaOrg JSON-LD file into URI
+    """
+    context = schema["@context"]
+    graph = schema["@graph"]
+    new_schema = {"@context": context,
+                  "@graph": [],
+                  "@id": schema["@id"]}
+    for record in graph:
+        new_record = {}
+        for k, v in record.items():
+            if type(v) == str:
+                new_record[expand_curie_to_uri(k, context)] =  expand_curie_to_uri(v, context)
+            elif type(v) == list:
+                if type(v[0]) == dict:
+                    new_record[expand_curie_to_uri(k, context)] = []
+                    for _item in v:
+                        new_record[expand_curie_to_uri(k, context)].append({"@id": expand_curie_to_uri(_item["@id"], context)})
+                else:
+                    new_record[expand_curie_to_uri(k, context)] = [expand_curie_to_uri(_item, context) for _item in v]
+            elif type(v) == dict and "@id" in v:
+                new_record[expand_curie_to_uri(k, context)] = {"@id": expand_curie_to_uri(v["@id"], context)}
+            elif v == None:
+                new_record[expand_curie_to_uri(k, context)] = None
+        new_schema["@graph"].append(new_record)
+    return new_schema
+
+
+def uri2label(uri, schema):
+    """Given a URI, return the label
+    """
+    return [record["rdfs:label"] for record in schema["@graph"] if record['@id'] == uri][0]
+
+
+def find_duplicates(_list):
+    """Find duplicate items in a list
+    """
+    return set([x for x in _list if _list.count(x) > 1])
